fix: cmake

+++ mvh.sh
@@ -29,7 +29,7 @@ cp -r "$SCRPIT_ROOT_DIR/build/libzhuziyi.a"  "$SCRPIT_ROOT_DIR/liautoinc/lib/lib
 clean_directory() {
     local dir="$1"
     for file in "$dir"/*; do
-        if [[ -f "$file" && "${file##*.}" != "h" ]]; then
+        if [[ -f "$file" && "${file##*.}" != "h" && "${file##*.}" != "inc" ]]; then
             rm -f "$file"
         elif [[ -d "$file" ]]; then
             clean_directory "$file"
